+++ src/main/java/de/mossgrabers/convertwithmoss/core/IMultisampleSource.java
@@ -82,6 +82,34 @@ public interface IMultisampleSource extends ISource
 
 
     /**
+     * If all zones of all groups are set to play round-robin.
+     *
+     * @return True if all sample zones have round-robin enabled
+     */
+    default boolean isFullRoundRobin ()
+    {
+        for (final IGroup group: this.getNonEmptyGroups (false))
+            if (!group.isFullRoundRobin ())
+                return false;
+        return true;
+    }
+
+
+    /**
+     * If at least one zones in this group is set to play round-robin.
+     *
+     * @return True if at least one sample zone has round-robin enabled
+     */
+    default boolean hasRoundRobin ()
+    {
+        for (final IGroup group: this.getNonEmptyGroups (false))
+            if (group.hasRoundRobin ())
+                return true;
+        return false;
+    }
+
+
+    /**
      * Collect all sample zones from all groups into one list.
      * 
      * @param filterReleaseTriggers Removes all groups which do only contain release triggers